Add PeerHandler
